R#18: loop wrap interpolation, VolEnv delay/hold, scaleTuning, SF2 send, sm24
1. Loop boundary wrap: interpolation taps near loopEnd now wrap to
   loopStart instead of clamping to sampleEnd. Eliminates clicks and
   buzzing on sustained instruments (organ, strings, synth pads).

2. VolEnv delay/hold stages: SF2 gen 25 (delayVolEnv) and gen 26
   (holdVolEnv) now implemented as separate envelope stages (delay→
   attack→hold→decay→sustain). Instruments with attack delay (e.g.
   some pad sounds) now have natural onset timing.

3. Scale tuning: SF2 gen 24 (scaleTuning, cents per key) now applied
   to pitch calculation. Non-equal temperament instruments (e.g.
   gamelan, fretless bass) render with correct intonation.

4. SF2 reverb/chorus send: SF2 gen 15/16 (reverbEffectsSend,
   chorusEffectsSend) parsed and stored per-zone for future per-
   voice FX routing.

5. sm24 24-bit sample support: parseSDTA now reads sm24 chunk and
   merges lower 8 bits into smpl 16-bit data for 24-bit resolution.
   sampleData type changed from int16_t to int32_t.

+++ cpp/include/soundfont.h
@@ -97,7 +97,7 @@ public:
     const std::vector<SF2Generator>& presetGenerators() const { return m_presetGenerators; }
     const std::vector<SF2InstBag>& instBags() const { return m_instBags; }
     const std::vector<SF2Generator>& instGenerators() const { return m_instGenerators; }
-    const int16_t* sampleData() const { return m_sampleData.data(); }
+    const int32_t* sampleData() const { return m_sampleData.data(); }
     size_t sampleDataSize() const { return m_sampleData.size(); }
 
     int findPreset(uint16_t bank, uint16_t preset) const;
@@ -108,7 +108,7 @@ private:
     bool parseSDTA(const uint8_t* data, size_t size);
     bool parsePDTA(const uint8_t* data, size_t size);
 
-    std::vector<int16_t> m_sampleData;
+    std::vector<int32_t> m_sampleData;
     std::vector<SF2Sample> m_samples;
     std::vector<SF2Preset> m_presets;
     std::vector<SF2PresetBag> m_presetBags;
